ctf/nsec/2018/captcha/requirements.txt
requests
opencv-python
ipython
numpy
pytesseract
Crypto
scipy
beautifulsoup4
pillow >= 2.0.0
captcha
captcha-solver